compile only linked .styl file
We tell Gulp to check only the main site.styl file that we import the others from
because otherwise it will tell Stylus to compile each file separately with
disastrous results. Many of the .styl files in the framework depend on others
being imported in the correct order by the framework’s index.styl.

NOTE: If you want to link to any separate compiled .css files in your
template, you need to also include them in the stylusSrc.