move template context vars to avoid errors
Adding function parameters to the context will cause startup errors if
they are added in the wrong way. Adding them the context in a router
middleware tends to cause startup errors where Pug finds that they are
undefined. Adding them to the context in an app middleware or koa-pug’s
global locals option doesn’t seem to cause problems.